+++ tools/serverpod_cli/test/integration/analyzer/dart/future_call_validation/future_call_method_test.dart
@@ -12,20 +12,23 @@ import '../../../../test_util/builders/generator_config_builder.dart';
 import '../../../../test_util/endpoint_validation_helpers.dart';
 
 var pathToServerpodRoot = Directory('../..').absolute.path;
-var testProjectDirectory = Directory.systemTemp.createTempSync('cli_test_');
+late Directory testProjectDirectory;
 
 void main() {
   setUpAll(() async {
+    testProjectDirectory = Directory.systemTemp.createTempSync('cli_test_');
     await createTestEnvironment(testProjectDirectory, pathToServerpodRoot);
   });
 
   tearDownAll(() {
-    testProjectDirectory.deleteSync(recursive: true);
+    if (testProjectDirectory.existsSync()) {
+      testProjectDirectory.deleteSync(recursive: true);
+    }
   });
 
   group('Given a valid future call with a method when analyzed', () {
     var collector = CodeGenerationCollector();
-    var testDirectory = Directory(
+    late final testDirectory = Directory(
       path.join(testProjectDirectory.path, const Uuid().v4()),
     );
 
@@ -95,7 +98,7 @@ class ExampleFutureCall extends FutureCall {
     'Given a valid future call method with a first positional nullable `Session` parameter when analyzed',
     () {
       var collector = CodeGenerationCollector();
-      var testDirectory = Directory(
+      late final testDirectory = Directory(
         path.join(testProjectDirectory.path, const Uuid().v4()),
       );
 
@@ -180,7 +183,7 @@ class ExampleFutureCall extends FutureCall {
     'of type `Session` when analyzed',
     () {
       var collector = CodeGenerationCollector();
-      var testDirectory = Directory(
+      late final testDirectory = Directory(
         path.join(testProjectDirectory.path, const Uuid().v4()),
       );
 
@@ -225,7 +228,7 @@ class ExampleFutureCall extends FutureCall {
     'Given a future call with a method that has a `Session` as required named parameter when analyzed',
     () {
       var collector = CodeGenerationCollector();
-      var testDirectory = Directory(
+      late final testDirectory = Directory(
         path.join(testProjectDirectory.path, const Uuid().v4()),
       );
 
@@ -272,7 +275,7 @@ class ExampleFutureCall extends FutureCall {
     'Given a future call method without a first positional `Session` parameter',
     () {
       var collector = CodeGenerationCollector();
-      var testDirectory = Directory(
+      late final testDirectory = Directory(
         path.join(testProjectDirectory.path, const Uuid().v4()),
       );
 
@@ -329,7 +332,7 @@ class ExampleFutureCall extends FutureCall {
     'and the first parameter instead contains a named `Session` parameter when analyzed',
     () {
       var collector = CodeGenerationCollector();
-      var testDirectory = Directory(
+      late final testDirectory = Directory(
         path.join(testProjectDirectory.path, const Uuid().v4()),
       );
 
@@ -386,7 +389,7 @@ class ExampleFutureCall extends FutureCall {
     'and the first parameter instead contains an optional `Session` parameter when analyzed',
     () {
       var collector = CodeGenerationCollector();
-      var testDirectory = Directory(
+      late final testDirectory = Directory(
         path.join(testProjectDirectory.path, const Uuid().v4()),
       );
 
@@ -442,7 +445,7 @@ class ExampleFutureCall extends FutureCall {
     'Given a future call method that without a Future<void> return type when analyzed',
     () {
       var collector = CodeGenerationCollector();
-      var testDirectory = Directory(
+      late final testDirectory = Directory(
         path.join(testProjectDirectory.path, const Uuid().v4()),
       );
 
@@ -504,7 +507,7 @@ class ExampleFutureCall extends FutureCall {
     'Given a future call method that returns a Future with non void type when analyzed',
     () {
       var collector = CodeGenerationCollector();
-      var testDirectory = Directory(
+      late final testDirectory = Directory(
         path.join(testProjectDirectory.path, const Uuid().v4()),
       );
 
@@ -564,7 +567,7 @@ class ExampleFutureCall extends FutureCall {
 
   group('Given a valid future call with private method when analyzed', () {
     var collector = CodeGenerationCollector();
-    var testDirectory = Directory(
+    late final testDirectory = Directory(
       path.join(testProjectDirectory.path, const Uuid().v4()),
     );
 
@@ -615,7 +618,7 @@ class ExampleFutureCall extends FutureCall {
     'Given a valid future call with multiple methods defined when analyzed',
     () {
       var collector = CodeGenerationCollector();
-      var testDirectory = Directory(
+      late final testDirectory = Directory(
         path.join(testProjectDirectory.path, const Uuid().v4()),
       );
 
@@ -677,7 +680,7 @@ class ExampleFutureCall extends FutureCall {
     'Given a valid future call method with documentation when analyzed',
     () {
       var collector = CodeGenerationCollector();
-      var testDirectory = Directory(
+      late final testDirectory = Directory(
         path.join(testProjectDirectory.path, const Uuid().v4()),
       );
 
@@ -737,7 +740,7 @@ class ExampleFutureCall extends FutureCall {
     'Given a valid future call method with "@Deprecated(<string literal>)" annotation',
     () {
       var collector = CodeGenerationCollector();
-      var testDirectory = Directory(
+      late final testDirectory = Directory(
         path.join(testProjectDirectory.path, const Uuid().v4()),
       );
 
@@ -797,7 +800,7 @@ class ExampleFutureCall extends FutureCall {
     'Given a valid future call method with "@Deprecated(<string const expr>)" annotation',
     () {
       var collector = CodeGenerationCollector();
-      var testDirectory = Directory(
+      late final testDirectory = Directory(
         path.join(testProjectDirectory.path, const Uuid().v4()),
       );
 
@@ -857,7 +860,7 @@ class ExampleFutureCall extends FutureCall {
 
   group('Given a valid future call method with "@deprecated" annotation', () {
     var collector = CodeGenerationCollector();
-    var testDirectory = Directory(
+    late final testDirectory = Directory(
       path.join(testProjectDirectory.path, const Uuid().v4()),
     );
 
@@ -916,7 +919,7 @@ class ExampleFutureCall extends FutureCall {
     'Given a valid future call with a method that has serializable parameters after the first positional Session parameter',
     () {
       var collector = CodeGenerationCollector();
-      var testDirectory = Directory(
+      late final testDirectory = Directory(
         path.join(testProjectDirectory.path, const Uuid().v4()),
       );
 
@@ -1002,7 +1005,7 @@ class ExampleFutureCall extends FutureCall {
     'Given a valid future call with a method that has a SerializableModel parameter after the Session parameter',
     () {
       var collector = CodeGenerationCollector();
-      var testDirectory = Directory(
+      late final testDirectory = Directory(
         path.join(testProjectDirectory.path, const Uuid().v4()),
       );
 
@@ -1067,7 +1070,7 @@ class ExampleFutureCall extends FutureCall {
     'Given a valid future call with a method that has non serializable parameters after the first positional Session parameter',
     () {
       var collector = CodeGenerationCollector();
-      var testDirectory = Directory(
+      late final testDirectory = Directory(
         path.join(testProjectDirectory.path, const Uuid().v4()),
       );
 
@@ -1129,7 +1132,7 @@ class ExampleFutureCall extends FutureCall {
     'Given a future call with a method that has a positional `Session` parameter only',
     () {
       var collector = CodeGenerationCollector();
-      var testDirectory = Directory(
+      late final testDirectory = Directory(
         path.join(testProjectDirectory.path, const Uuid().v4()),
       );
